Fixed review comment
