--- src/IceRpc/Slice/Payload.cs
@@ -1,113 +1,0 @@
-// Copyright (c) ZeroC, Inc. All rights reserved.
-
-using IceRpc.Slice.Internal;
-using System.Diagnostics;
-
-namespace IceRpc.Slice
-{
-    /// <summary>Methods to read and write the payloads of requests and responses.</summary>
-    public static class Payload
-    {
-        /// <summary>Creates the payload of a request from the request's arguments. Use this method is for operations
-        /// with multiple parameters.</summary>
-        /// <typeparam name="T">The type of the operation's parameters.</typeparam>
-        /// <param name="proxy">A proxy to the target service.</param>
-        /// <param name="args">The arguments to write into the payload.</param>
-        /// <param name="encodeAction">The <see cref="TupleEncodeAction{T}"/> that encodes the arguments into the
-        /// payload.</param>
-        /// <param name="classFormat">The class format in case any parameter is a class.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromArgs<T>(
-            this Proxy proxy,
-            in T args,
-            TupleEncodeAction<T> encodeAction,
-            FormatType classFormat = default) where T : struct
-        {
-            var bufferWriter = new BufferWriter();
-            var encoder = proxy.Encoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
-            encodeAction(encoder, in args);
-            return bufferWriter.Finish();
-        }
-
-        /// <summary>Creates the payload of a request without parameter.</summary>
-        /// <param name="proxy">A proxy to the target service.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromEmptyArgs(this Proxy proxy)
-        {
-            proxy.Encoding.CheckSupportedIceEncoding();
-            return default;
-        }
-
-        /// <summary>Creates the payload of a response from the request's dispatch and return value tuple. Use this
-        /// method when the operation returns a tuple.</summary>
-        /// <typeparam name="T">The type of the operation's return value tuple.</typeparam>
-        /// <param name="payloadEncoding">The payload encoding.</param>
-        /// <param name="returnValueTuple">The return values to write into the payload.</param>
-        /// <param name="encodeAction">The <see cref="TupleEncodeAction{T}"/> that encodes the arguments into the
-        /// payload.</param>
-        /// <param name="classFormat">The class format in case T is a class.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromReturnValueTuple<T>(
-            this Encoding payloadEncoding,
-            in T returnValueTuple,
-            TupleEncodeAction<T> encodeAction,
-            FormatType classFormat = default) where T : struct
-        {
-            var bufferWriter = new BufferWriter();
-            var encoder = payloadEncoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
-            encodeAction(encoder, in returnValueTuple);
-            return bufferWriter.Finish();
-        }
-
-        /// <summary>Creates the payload of a request from the request's argument. Use this method when the operation
-        /// takes a single parameter.</summary>
-        /// <typeparam name="T">The type of the operation's parameter.</typeparam>
-        /// <param name="proxy">A proxy to the target service.</param>
-        /// <param name="arg">The argument to write into the payload.</param>
-        /// <param name="encodeAction">The <see cref="EncodeAction{T}"/> that encodes the argument into the payload.
-        /// </param>
-        /// <param name="classFormat">The class format in case T is a class.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromSingleArg<T>(
-            this Proxy proxy,
-            T arg,
-            EncodeAction<T> encodeAction,
-            FormatType classFormat = default)
-        {
-            var bufferWriter = new BufferWriter();
-            var encoder = proxy.Encoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
-            encodeAction(encoder, arg);
-            return bufferWriter.Finish();
-        }
-
-        /// <summary>Creates the payload of a response from the request's dispatch and return value. Use this method
-        /// when the operation returns a single value.</summary>
-        /// <typeparam name="T">The type of the operation's parameter.</typeparam>
-        /// <param name="payloadEncoding">The payload encoding.</param>
-        /// <param name="returnValue">The return value to write into the payload.</param>
-        /// <param name="encodeAction">The <see cref="EncodeAction{T}"/> that encodes the argument into the payload.
-        /// </param>
-        /// <param name="classFormat">The class format in case T is a class.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromSingleReturnValue<T>(
-            this Encoding payloadEncoding,
-            T returnValue,
-            EncodeAction<T> encodeAction,
-            FormatType classFormat = default)
-        {
-            var bufferWriter = new BufferWriter();
-            var encoder = payloadEncoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
-            encodeAction(encoder, returnValue);
-            return bufferWriter.Finish();
-        }
-
-        /// <summary>Creates a payload representing a void return value.</summary>
-        /// <param name="payloadEncoding">The payload encoding.</param>
-        /// <returns>A new payload.</returns>
-        public static ReadOnlyMemory<ReadOnlyMemory<byte>> FromVoidReturnValue(this Encoding payloadEncoding)
-        {
-            payloadEncoding.CheckSupportedIceEncoding();
-            return default;
-        }
-    }
-}
