+++ src/IceRpc/Slice/ProxyExtensions.cs
@@ -1,5 +1,7 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
+using IceRpc.Slice.Internal;
+
 namespace IceRpc.Slice
 {
     /// <summary>A function that decodes the return value from an Ice-encoded response.</summary>
@@ -17,6 +19,57 @@ namespace IceRpc.Slice
     /// <summary>Provides extension methods for class Proxy.</summary>
     public static class ProxyExtensions
     {
+        /// <summary>Creates the payload of a request without parameter.</summary>
+        /// <param name="proxy">A proxy to the target service.</param>
+        /// <returns>A new payload.</returns>
+        public static ReadOnlyMemory<ReadOnlyMemory<byte>> CreateEmptyPayload(this Proxy proxy)
+        {
+            proxy.Encoding.CheckSupportedIceEncoding();
+            return default;
+        }
+
+        /// <summary>Creates the payload of a request from the request's argument. Use this method when the operation
+        /// takes a single parameter.</summary>
+        /// <typeparam name="T">The type of the operation's parameter.</typeparam>
+        /// <param name="proxy">A proxy to the target service.</param>
+        /// <param name="arg">The argument to write into the payload.</param>
+        /// <param name="encodeAction">The <see cref="EncodeAction{T}"/> that encodes the argument into the payload.
+        /// </param>
+        /// <param name="classFormat">The class format in case T is a class.</param>
+        /// <returns>A new payload.</returns>
+        public static ReadOnlyMemory<ReadOnlyMemory<byte>> CreatePayloadFromSingleArg<T>(
+            this Proxy proxy,
+            T arg,
+            EncodeAction<T> encodeAction,
+            FormatType classFormat = default)
+        {
+            var bufferWriter = new BufferWriter();
+            var encoder = proxy.Encoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
+            encodeAction(encoder, arg);
+            return bufferWriter.Finish();
+        }
+
+        /// <summary>Creates the payload of a request from the request's arguments. Use this method is for operations
+        /// with multiple parameters.</summary>
+        /// <typeparam name="T">The type of the operation's parameters.</typeparam>
+        /// <param name="proxy">A proxy to the target service.</param>
+        /// <param name="args">The arguments to write into the payload.</param>
+        /// <param name="encodeAction">The <see cref="TupleEncodeAction{T}"/> that encodes the arguments into the
+        /// payload.</param>
+        /// <param name="classFormat">The class format in case any parameter is a class.</param>
+        /// <returns>A new payload.</returns>
+        public static ReadOnlyMemory<ReadOnlyMemory<byte>> CreatePayloadFromArgs<T>(
+            this Proxy proxy,
+            in T args,
+            TupleEncodeAction<T> encodeAction,
+            FormatType classFormat = default) where T : struct
+        {
+            var bufferWriter = new BufferWriter();
+            var encoder = proxy.Encoding.CreateIceEncoder(bufferWriter, classFormat: classFormat);
+            encodeAction(encoder, in args);
+            return bufferWriter.Finish();
+        }
+
         /// <summary>Sends a request to a service and decodes the response.</summary>
         /// <param name="proxy">A proxy for the remote service.</param>
         /// <param name="operation">The name of the operation, as specified in Slice.</param>
